requirements.txt
sentence-transformers>=2.2.2
chromadb>=1.0.0
chromadb-client>=1.0.0
rank_bm25>=0.2.2
nltk>=3.8.1
transformers>=4.36.2
litellm>=1.18.5
numpy>=1.24.3
scikit-learn>=1.3.2
openai>=1.3.7
pydantic>=2.6.3
python-dotenv>=1.0.0
pytest>=7.4.0
requests>=2.31.0
einops>=0.7.0

# Google Gemini dependencies
google-generativeai>=0.3.2
langchain>=0.0.329
langchain-google-genai>=0.0.7

# MCP (Model Context Protocol) dependencies
mcp[cli]>=1.6.0

# ASGI Server
uvicorn>=0.34.0

# WebSocket Client
websockets>=15.0.1

# HTTP Client (for SSE)
httpx>=0.20.0

# ASGI Framework (for mounting sse_app)
starlette>=0.40.0,<0.46.0
sse-starlette>=2.2.1

colorlog>=6.9.0

psutil>=7.0.0
